Fix unit tests after health endpoint changes
