+++ tests/test_auth.py
@@ -4,6 +4,7 @@
 import unittest.mock
 import hmac
 from types import SimpleNamespace
+import time
 
 
 import aiohttp.web
@@ -46,7 +47,8 @@ class AuthModuleTestCase(asynctest.TestCase):
             await t_signature(
                 tokens_mock,
                 "",
-                message_mock
+                message_mock,
+                time.time() + 3600,
             )
 
     async def test_test_signature_success(self):
@@ -61,7 +63,8 @@ class AuthModuleTestCase(asynctest.TestCase):
         await t_signature(
             tokens_mock,
             signature,
-            message_mock
+            message_mock,
+            time.time() + 3600
         )
 
     async def test_handle_validate_authentication_success(self):
@@ -83,7 +86,9 @@ class AuthModuleTestCase(asynctest.TestCase):
             },
             "url": SimpleNamespace(**{
                 "path": "c"
-            })
+            }),
+            "match_info": {},
+            "path": "/health",
         })
 
         with t_signature_patch:
@@ -102,7 +107,10 @@ class AuthModuleTestCase(asynctest.TestCase):
             },
             "url": SimpleNamespace(**{
                 "path": "c"
-            })
+            }),
+            "match_info": {},
+            "path": "/health",
+
         })
         with self.assertRaises(aiohttp.web.HTTPClientError):
             await handle_validate_authentication(
